+++ src/main/java/org/camunda/webapptranslation/report/ReportInt.java
@@ -1,4 +1,4 @@
-package org.camunda.webapptranslation.org.camunda.webapptranslation;
+package org.camunda.webapptranslation.report;
 
 public interface ReportInt {
